# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Pistas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
        "/api/add-warning/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Añadir una advertencia a un usuario",
                "operationId": "6569f007ffbc9694473b25171a107faf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "adv",
                        "in": "header",
                        "description": "El texto de advertencia para el usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Operación realizada con éxito\n    *           Primera advertencia añadida con éxito\n    *           Segunda advertencia añadida con éxito. La cuenta ha sido bloqueada"
                    },
                    "400": {
                        "description": "\n    *           No se encuentra el usuario\n    *           Error de validacion"
                    },
                    "403": {
                        "description": "\n    *           Esta operación solo lo puede hacer un administrador o un moderador"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-role/{id}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtener el rol del usuario por su id",
                "operationId": "a58383bb077c321c0988c770515cf43d",
                "responses": {
                    "202": {
                        "description": "\n    *           $rol_id (integer)"
                    },
                    "400": {
                        "description": "\n    *           $Este usuario no existe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-warnings/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtener las advertencias del usuario puestas por algún Moderador / Administrador",
                "operationId": "b5aca3d7f81c52eeaefbad6af746d65d",
                "responses": {
                    "202": {
                        "description": "\n    *           $adv1 (string)\n    *           $adv2 (string)"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtener todos los usuarios",
                "operationId": "36d93c62e98bba46af2f67392fb0b2b3",
                "responses": {
                    "202": {
                        "description": "\n    *           Se han obtenido los usuarios correctamente\n    *           $users (Object)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener usuario y ver información de este",
                "operationId": "8676e42cf7da7995603730250593a5db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           $user (Object)"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtener usuario y ver información de este",
                "operationId": "6061bd3baaa2141c0499ae28e3d36426",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           $user (Object)"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    },
                    "403": {
                        "description": "\n    *           Esta operación solo lo puede hacer un administrador o un moderador"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-account/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Borrar cuenta de un usuario",
                "operationId": "cc62c09cd4f72796834e9c4b2d6eca76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha eliminado la cuenta correctamente"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    },
                    "403": {
                        "description": "\n    *           Esta operación solo lo puede hacer un administrador"
                    },
                    "406": {
                        "description": "\n    *           No se ha podido eliminar la cuenta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/edit-user/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Actualizar nombre de un usuario",
                "operationId": "948833be8312920f0c6984432b8ffd36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha actualizado correctamente el nombre de usuario"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    },
                    "403": {
                        "description": "\n    *           Esta operación solo lo puede hacer un administrador"
                    },
                    "406": {
                        "description": "\n    *           No se ha podido actualizar el nombre de usuario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/edit-email/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Actualizar correo electrónico de un usuario",
                "operationId": "165cd8ef15ce443c065e336272879fc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha actualizado correctamente el email del usuario"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    },
                    "403": {
                        "description": "\n    *           Esta operación solo lo puede hacer un administrador"
                    },
                    "406": {
                        "description": "\n    *           No se ha podido actualizar correctamente el email del usuario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}